+++ src/RewardFlow.API/Rewards/Data/Subject.cs
@@ -9,7 +9,7 @@ public class Subject : ITenantEntity
     public string Name { get; set; } = null!;
     public bool IsTheoretical { get; set; }
     public bool IsPractical { get; set; }
-    public float SubjectPrice { get; set; }
+    public decimal SubjectPrice { get; set; }
     
     public virtual ICollection<SemesterSubject> SubjectSemesters { get; set; } = new List<SemesterSubject>();
 }
